requirements.txt
asgiref==3.5.2
Django==4.1.3
django-filter==22.1
django-registration-redux==1.2
django-rest-registration==0.7.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.0.4
generic==1.1.0
importlib-metadata==5.0.0
Markdown==3.4.1
psycopg2==2.9.5
Pygments==2.13.0
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
zipp==3.10.0
